moved registry files into registries folder
