+++ src/Sulu/Bundle/AdminBundle/Resources/js/containers/Datagrid/index.js
@@ -1,7 +1,7 @@
 // @flow
 import Datagrid from './Datagrid';
 import DatagridStore from './stores/DatagridStore';
-import datagridAdapterRegistry from './DatagridAdapterRegistry';
+import datagridAdapterRegistry from './registries/DatagridAdapterRegistry';
 import TableAdapter from './adapters/TableAdapter';
 import FolderAdapter from './adapters/FolderAdapter';
 import type {DatagridAdapter, DatagridAdapterProps} from './types';
