+++ src/Sulu/Bundle/AdminBundle/Resources/js/containers/Datagrid/registries/DatagridAdapterRegistry.js
@@ -1,5 +1,5 @@
 // @flow
-import type {DatagridAdapter} from './types';
+import type {DatagridAdapter} from '../types';
 
 class DatagridAdapterRegistry {
     adapters: {[string]: DatagridAdapter};
