+++ src/Sulu/Bundle/AdminBundle/Resources/js/containers/Form/index.js
@@ -1,6 +1,6 @@
 // @flow
 import Form from './Form';
-import fieldRegistry from './FieldRegistry';
+import fieldRegistry from './registries/FieldRegistry';
 
 export {fieldRegistry};
 export default Form;
