+++ src/Sulu/Bundle/AdminBundle/Resources/js/containers/Form/registries/FieldRegistry.js
