+++ src/Sulu/Bundle/AdminBundle/Resources/js/containers/ViewRenderer/registries/ViewRegistry.js
@@ -1,5 +1,5 @@
 // @flow
-import type {View} from './types';
+import type {View} from '../types';
 
 class ViewRegistry {
     views: {[string]: View};
