+++ src/Sulu/Bundle/AdminBundle/Resources/js/containers/ViewRenderer/tests/ViewRegistry.test.js
@@ -1,6 +1,6 @@
 /* eslint-disable flowtype/require-valid-file-annotation */
 import React from 'react';
-import viewRegistry from '../ViewRegistry';
+import viewRegistry from '../registries/ViewRegistry';
 
 beforeEach(() => {
     viewRegistry.clear();
